codewide: delete internal APIs & impls
This huge commit deletes most of the internal APIs from the codebase.
This is an important basis for reimplementation of the public APIs
as a wrapper over Rust.

Rationale:
Even if subset of the deleted internal APIs could prove useful in the
new implementation, we prefer to start from the cleanest version
possible and then add new code, possibly reintroducing parts of the
previously deleted code. This will:
- facilitate understanding of the existing public APIs,
- reduce disturbance arising from left-over private implementations,
- make us less prone to accidentally keep too much legacy internal code
  from the previous implementation.